% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_definitions.R
\docType{data}
\name{munch_colors}
\alias{munch_colors}
\title{Color palette options based on individual paintings}
\format{
An object of class \code{list} of length 13.
}
\usage{
munch_colors
}
\description{
To construct a palette, \code{\link{munch_palette}} and specify one of the options.
}
\keyword{datasets}
