% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_definitions.R
\name{munch_palette}
\alias{munch_palette}
\title{Function that generates colors based on palette choice and number of colors.}
\usage{
munch_palette(choice, n, discrete = FALSE)
}
\arguments{
\item{choice}{Name of color palette desired.}

\item{n}{Number of colors in the palette desired. Palettes have between four and
eleven colors. If discrete is specified as TRUE, then n is limited by the number
of colors in the original palette.}

\item{discrete}{If the palette is to be discrete (TRUE) or continuous (FALSE).
Default option is FALSE.}
}
\value{
A vector of colors in hex format.
}
\description{
Function that generates colors based on palette choice and number of colors.
}
\examples{
munch_palette("StarryNight1", 100)
munch_palette("NewSnowCool", 4, TRUE)
}
\keyword{colors}
