% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munchcolors.R
\docType{package}
\name{munchcolors}
\alias{munchcolors}
\title{munchcolors}
\description{
The munchcolors package contains a set of color palettes and a function
   for generating color palettes of different sizes based on the original palette
   options.
}
\details{
A package of color palettes based on Munch paintings
}
