% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_definitions.R
\name{scale_fill_munch}
\alias{scale_fill_munch}
\title{Function that generates fill colors for a ggplot based on palette choice and number of colors.}
\usage{
scale_fill_munch(..., choice, discrete = FALSE)
}
\arguments{
\item{...}{Parameters passed to discrete_scale if discrete == TRUE or scale_fill_gradientn if FALSE.}

\item{choice}{Name of color palette desired.}

\item{discrete}{If the palette is to be discrete (TRUE) or continuous (FALSE).
Default option is FALSE.}
}
\description{
Function that generates fill colors for a ggplot based on palette choice and number of colors.
}
\keyword{colors}
\keyword{ggplot}
